Added asynchronous scheduler and simple test to it
