+++ requirements-dev.txt
@@ -1,3 +1,4 @@
+aiounittest
 docutils
 mock
 Pygments
